+++ be/src/main/java/com/niels/homebanking/dto/currency/UpdateCurrencyDto.java
@@ -1,5 +1,7 @@
 package com.niels.homebanking.dto.currency;
 
+import java.math.BigDecimal;
+
 public class UpdateCurrencyDto {
 
     private String name;
@@ -8,12 +10,12 @@ public class UpdateCurrencyDto {
 
     private String symbol;
 
-    private long exchange;
+    private BigDecimal exchange;
 
     public UpdateCurrencyDto() {
     }
 
-    public UpdateCurrencyDto(String name, String iso, String symbol, long exchange) {
+    public UpdateCurrencyDto(String name, String iso, String symbol, BigDecimal exchange) {
         this.name = name;
         this.iso = iso;
         this.symbol = symbol;
@@ -44,11 +46,11 @@ public class UpdateCurrencyDto {
         this.symbol = symbol;
     }
 
-    public long getExchange() {
+    public BigDecimal getExchange() {
         return exchange;
     }
 
-    public void setExchange(long exchange) {
+    public void setExchange(BigDecimal exchange) {
         this.exchange = exchange;
     }
 }
